SensoryCloud/src/main/java/ai/sensorycloud/tokenManager/TokenManager.java
package ai.sensorycloud.tokenManager;

import ai.sensorycloud.api.common.TokenResponse;
import ai.sensorycloud.service.OAuthService;
import io.grpc.ClientInterceptor;
import io.grpc.Context;
import io.grpc.Metadata;
import io.grpc.stub.MetadataUtils;

import java.util.concurrent.locks.ReentrantLock;

/**
 * A token manager that manages storing and fetching OAuth tokens for Sensory Cloud
 */
public class TokenManager {

    private String token = "";
    private Long expires = 0L;

    private final long expirationBuffer = 5 * 60 * 1000; // 5 minutes in ms

    private Context context;
    private OAuthService oAuthService;
    private ReentrantLock tokenMutex = new ReentrantLock();

    /**
     * Creates a new Token Manager instance
     *
     * @param context Application context, used for caching the current OAuth token
     * @param oAuthService OAuth service for fetching new OAuth tokens from
     */
    public TokenManager(Context context, OAuthService oAuthService) {
        this.context = context;
        this.oAuthService = oAuthService;
    }

    /**
     * Returns a valid OAuth access token
     * This call will fetch a new token if the current token is expired
     *
     * @return A valid OAuth token
     * @Throws Exception – if an error occurs while requesting a new token
     */
    public String getAccessToken() throws Exception {
        this.tokenMutex.lock();
        long now = System.currentTimeMillis();
        String token = this.token;
        if (this.token.equals("") || now > this.expires - expirationBuffer) {
            try {
                token = fetchNewAccessToken();
            } catch (Exception ex) {
                this.tokenMutex.unlock();
                throw  ex;
            }
        }

        this.tokenMutex.unlock();
        return token;
    }

    /**
     * Returns an OAuth authorization header
     * This call will fetch a new token if the current token is expired
     *
     * @return A valid OAuth authorization header
     * @Throws Exception – if an error occurs while requesting a new token
     */
    public ClientInterceptor getAuthorizationMetadata() throws Exception {
        String token = getAccessToken();
        Metadata header = new Metadata();
        Metadata.Key<String> key = Metadata.Key.of("authorization", Metadata.ASCII_STRING_MARSHALLER);
        header.put(key, "Bearer " + token);
        return MetadataUtils.newAttachHeadersInterceptor(header);
    }

    /**
     * Deletes the currently cached OAuth token
     */
    public void deleteCachedToken() {
        this.token = "";
        this.expires = 0L;
    }

    private String fetchNewAccessToken() throws Exception {
        TokenResponse response = oAuthService.getTokenSync();
        this.token = response.getAccessToken();
        this.expires = System.currentTimeMillis() + (response.getExpiresIn() * 1000L);
        return response.getAccessToken();
    }
}
